#29 Centralized file for auth typedefs and exceptions.
